requirements.txt
requests
pyyaml
SQLAlchemy
psycopg2-binary
openpyxl
schedule
asyncio
python-dotenv